move implementation of metadata query
